[rename] 폴더명, 파일 위치를 map브랜치 기준으로 일부 변경
