--- src/util/common/database.ts
@@ -1,16 +1,0 @@
-import { MongoClient } from "mongodb";
-
-const url = process.env.MONGODB_URL || "";
-const options = {};
-
-let connectDB: Promise<MongoClient>;
-
-if (process.env.NODE_ENV === "development") {
-  if (!global._mongo) {
-    global._mongo = new MongoClient(url, options).connect();
-  }
-  connectDB = global._mongo;
-} else {
-  connectDB = new MongoClient(url, options).connect();
-}
-export { connectDB };
